+++ src/mcast/ht/util/ManagementProperties.java
@@ -7,17 +7,17 @@ public class ManagementProperties {
 
 	private static final String MPROP_MESSAGE_BYTES = "MessageBytes";
 	
-	public static int getMessageBytes(Managable m) {
+	public static long getMessageBytes(Managable m) {
 		try {
 			String bytes = m.getManagementProperty(MPROP_MESSAGE_BYTES);
-			return Integer.parseInt(bytes);
+			return Long.parseLong(bytes);
 		} catch (NoSuchPropertyException e) {
 			throw new RuntimeException("Management property '" + 
 					MPROP_MESSAGE_BYTES + " cannot be read", e);
 		}
 	}
 	
-	public static void setMessageBytes(Managable m, int value) {
+	public static void setMessageBytes(Managable m, long value) {
 		try {
 			m.setManagementProperty(MPROP_MESSAGE_BYTES, String.valueOf(value));
 		} catch (NoSuchPropertyException e) {
